Restructures project. Adds setup.
