+++ .gitignore
@@ -7,3 +7,5 @@ out.jpg
 .ipynb_checkpoints/
 datasets
 winners/
+.idea
+.DS_Store
