+++ helper/group_by_aspect_ratio.py
